x86-64 support is now mostly functional with some fixes and improvements. * LowLevelHook is now functional in x64. * Using state-bound functors as hook callbacks is now functional in x64. * Fix dozens of type narrowing build warnings on x64.
Note that EditExports is still unsupported in x64. There is some code cleanup that needs doing as well.

+++ Patcher.h
@@ -268,7 +268,7 @@ public:
 
   PatcherStatus ReleaseModule();  ///< If this PatchContext has loaded a module, releases its active handle to it.
 
-  uint32 NumPatches() const { return history_.size(); }  ///< Returns the number of active patches.
+  size_t NumPatches() const { return history_.size(); }  ///< Returns the number of active patches.
 
   /// Returns true if the given address had previously been touched or patched.
   bool HasPatched(TargetPtr pAddress) const { return historyAt_.count(MaybeFixTargetPtr(pAddress)) != 0; }
